requirements.txt
advanced_alchemy==0.7.4
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
deprecation==2.1.0
distlib==0.3.8
ecdsa==0.19.0
fastapi==0.109.1
fastapi_keycloak==1.0.10
filelock==3.13.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
identify==2.5.33
idna==3.6
itsdangerous==2.1.2
Mako==1.3.2
MarkupSafe==2.1.5
nodeenv==1.8.0
packaging==23.2
passlib==1.7.4
platformdirs==4.1.0
pre-commit==3.6.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9
setuptools==70.1.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.35.1
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.26.0
uvloop==0.19.0
virtualenv==20.25.0
watchfiles==0.21.0
websockets==12.0
